+++ build_bindgen.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -euxo pipefail
-
-cargo build --release --package uniffi-bindgen-cpp