+++ cpp-tests/scaffolding_tests/custom_types_builtin/lib_custom_types_builtin.cpp
@@ -3,12 +3,17 @@
 #include <limits>
 
 custom_types_builtin::CustomTypesBuiltin custom_types_builtin::get_custom_types_builtin() {
-    return custom_types_builtin::CustomTypesBuiltin{
+    auto custom_string = custom_types_builtin::CustomString{"Custom string"};
+    auto array = std::vector<std::string>{"Hello, world!"};
+    auto bytes = std::vector<uint8_t>{0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
+    auto table = std::unordered_map<std::string, std::string>{{"hello", "world"}};
+
+    auto result = custom_types_builtin::CustomTypesBuiltin{
         .string = "Hello, world!",
-        .custom_string = custom_types_builtin::CustomString{"Custom string"},
-        .array = std::vector<std::string>{"Hello, world!"},
-        .bytes = std::vector<uint8_t>{0, 1, 2, 3, 4, 5, 6, 7, 8, 9},
-        .table = std::unordered_map<std::string, std::string>{{"hello", "world"}},
+        .custom_string = custom_string,
+        .array = array,
+        .bytes = bytes,
+        .table = table,
         .boolean = true,
         .int8 = std::numeric_limits<int8_t>::max(),
         .int16 = std::numeric_limits<int16_t>::max(),
@@ -19,8 +24,10 @@ custom_types_builtin::CustomTypesBuiltin custom_types_builtin::get_custom_types_
         .uint32 = std::numeric_limits<uint32_t>::max(),
         .uint64 = std::numeric_limits<uint64_t>::max(),
         .flt = std::numeric_limits<float>::max(),
-        .dbl = std::numeric_limits<double>::max(),
+        .dbl = std::numeric_limits<double>::max()
     };
+
+    return result;
 }
 
 custom_types_builtin::CustomTypesBuiltin custom_types_builtin::return_custom_types_builtin(custom_types_builtin::CustomTypesBuiltin type) {